LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/manual/draw-generic-family/2d.text.draw.generic.family.w-expected.txt

CONSOLE MESSAGE: Error: assert_array_equals: The image data generated by drawing generic font family 'sans-serif' should be the same for both OffscreenCanvas and regular canvas expected property 11 to be 0 but got 75 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,0,0,0,75,0,0,0,248,0,0,0,255,0")
CONSOLE MESSAGE: Error: assert_array_equals: The image data generated by drawing generic font family 'serif' should be the same for both OffscreenCanvas and regular canvas expected property 23 to be 0 but got 7 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,0,0,0,7,0,0,0,14,0,0,0,1,0")
CONSOLE MESSAGE: Error: assert_array_equals: The image data generated by drawing generic font family 'cursive' should be the same for both OffscreenCanvas and regular canvas expected property 31 to be 0 but got 5 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,0,0,0,5,0,0,0,15,0,0,0,3,0")
CONSOLE MESSAGE: Error: assert_array_equals: The image data generated by drawing generic font family 'fantasy' should be the same for both OffscreenCanvas and regular canvas expected property 15 to be 0 but got 123 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,0,0,0,123,0,0,0,255,0,0,0,255,0")
CONSOLE MESSAGE: Error: assert_array_equals: The image data generated by drawing generic font family 'monospace' should be the same for both OffscreenCanvas and regular canvas expected property 7 to be 0 but got 97 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,97,0,0,0,255,0,0,0,240,0,0,0,98")

Harness Error (FAIL), message = Error: assert_array_equals: The image data generated by drawing generic font family 'monospace' should be the same for both OffscreenCanvas and regular canvas expected property 7 to be 0 but got 97 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,97,0,0,0,255,0,0,0,240,0,0,0,98")

TIMEOUT Test that drawing sans-serif produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing serif produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing cursive produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing fantasy produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing monospace produces the same result between canvas and OffscreenCanvas in a Worker Test timed out

Harness Error (FAIL), message = Error: assert_array_equals: The image data generated by drawing generic font family 'monospace' should be the same for both OffscreenCanvas and regular canvas expected property 7 to be 0 but got 97 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "0,0,0,0,0,0,0,97,0,0,0,255,0,0,0,240,0,0,0,98")

TIMEOUT Test that drawing sans-serif produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing serif produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing cursive produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing fantasy produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
TIMEOUT Test that drawing monospace produces the same result between canvas and OffscreenCanvas in a Worker Test timed out
